// service/protos/servicev2_pb.d.ts
  export type AsObject = {
    playerId: number,
    gameId: number,
    connect?: ConnectMessage.AsObject,
    step?: StepMessage.AsObject,
    reset?: ResetMessage.AsObject,
  }
// ...
export class Action extends jspb.Message {
  getRqid(): number;
  setRqid(value: number): void;

  getValue(): number;
  setValue(value: number): void;

  getText(): string;
  setText(value: string): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): Action.AsObject;
  static toObject(includeInstance: boolean, msg: Action): Action.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: Action, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): Action;
  static deserializeBinaryFromReader(message: Action, reader: jspb.BinaryReader): Action;
}

export namespace Action {
  export type AsObject = {
    rqid: number,
    value: number,
    text: string,
  }
}

export class StepMessage extends jspb.Message {
  hasAction(): boolean;
  clearAction(): void;
  getAction(): Action | undefined;
  setAction(value?: Action): void;

  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): StepMessage.AsObject;
  static toObject(includeInstance: boolean, msg: StepMessage): StepMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: StepMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): StepMessage;
  static deserializeBinaryFromReader(message: StepMessage, reader: jspb.BinaryReader): StepMessage;
}

export namespace StepMessage {
  export type AsObject = {
    action?: Action.AsObject,
  }
}

export class ResetMessage extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ResetMessage.AsObject;
  static toObject(includeInstance: boolean, msg: ResetMessage): ResetMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: ResetMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ResetMessage;
  static deserializeBinaryFromReader(message: ResetMessage, reader: jspb.BinaryReader): ResetMessage;
}

export namespace ResetMessage {
  export type AsObject = {
  }
}

export class ConnectMessage extends jspb.Message {
  serializeBinary(): Uint8Array;
  toObject(includeInstance?: boolean): ConnectMessage.AsObject;
  static toObject(includeInstance: boolean, msg: ConnectMessage): ConnectMessage.AsObject;
  static extensions: {[key: number]: jspb.ExtensionFieldInfo<jspb.Message>};
  static extensionsBinary: {[key: number]: jspb.ExtensionFieldBinaryInfo<jspb.Message>};
  static serializeBinaryToWriter(message: ConnectMessage, writer: jspb.BinaryWriter): void;
  static deserializeBinary(bytes: Uint8Array): ConnectMessage;
  static deserializeBinaryFromReader(message: ConnectMessage, reader: jspb.BinaryReader): ConnectMessage;
}

export namespace ConnectMessage {
  export type AsObject = {
  }
}